+++ examples/quart/app/api/v1/__init__.py
@@ -1,0 +1,8 @@
+from fastopenapi.routers.quart import QuartRouter
+
+from .authors import router as authors_router
+from .posts import router as posts_router
+
+v1_router = QuartRouter()
+v1_router.include_router(authors_router)
+v1_router.include_router(posts_router)